# src/main/resources/swagger/Demo-API-1.0.0.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: DemoAPI
  description:
    DemoAPI.
  contact:
    name: Ali Ezat
    email: ali.hely@gmail.com
    url: 'http://gitlab-url'
  license:
    name: Copyright 2020
host: localhost
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /demo/accounts/current:
    post:
      summary:
        Open current account for existing customer.
      description:
        Accepts the user information (customerID, initialCredit) to open a new account connected to the user whose ID is customerID.
        If initialCredit is not 0, a transaction will be sent to the new account.
      parameters:
        - in: body
          name: openCurrentAccountRequest
          description:
            open current account for existing cutomer request.
          required: true
          schema:
            $ref: '#/definitions/openCurrentAccountRequest'
      responses:
        '200':
          description: Account created successfully.
          schema:
            $ref: '#/definitions/openCurrentAccountResponse'
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/errorObject'
        '403':
          description: Unauthorized request
          schema:
            $ref: '#/definitions/errorObject'
        '404':
          description: Not Found - The resource referenced in the request not found
          schema:
            $ref: '#/definitions/errorObject'
        '500':
          description: Technical server error
          schema:
            $ref: '#/definitions/errorObject'
        default:
          description: Unexpected unknown error!
          schema:
            $ref: '#/definitions/errorObject'
  /demo/customer/accounts/{customerId}:
    get:
      summary:
        get cutomer details and his accounts details.
      description:
        Accepts cutomer Id to reterive cutomer detials (name, surname), also reterive accounts details (account no and balance ) assigned to this customer.
      parameters:
        - in: path
          name: customerId
          type: string
          required: true
      responses:
        '200':
          description: cutomer details retrived successfully.
          schema:
            $ref: '#/definitions/customerDetailsResponse'
        '400':
          description: Invalid request
          schema:
            $ref: '#/definitions/errorObject'
        '403':
          description: Unauthorized request
          schema:
            $ref: '#/definitions/errorObject'
        '404':
          description: Not Found - The resource referenced in the request not found
          schema:
            $ref: '#/definitions/errorObject'
        '500':
          description: Technical server error
          schema:
            $ref: '#/definitions/errorObject'
        default:
          description: Unexpected unknown error!
          schema:
            $ref: '#/definitions/errorObject'
definitions:
  openCurrentAccountRequest:
    type: object
    required:
      - customerID
      - initialCredit
    properties:
      customerID:
        type: string
        description: Customer ID.
      initialCredit:
        type: number
        description: Initial credit for the new account.
        minimum: 0
        exclusiveMinimum: true
  openCurrentAccountResponse:
    type: object
    properties:
      message:
        type: string
      customerID:
        type: string
      newCurrentAccountID:
        type: string
  errorObject:
    type: object
    properties:
      code:
        type: string
      message:
        type: string
      target:
        type: string
      details:
        type: array
        items:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            target:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  target:
                    type: string
  customerDetailsResponse:
    type: object
    required:
      - customerDetails
    properties:
      customer:
        type: object
        $ref: '#/definitions/customerDetails'
  customerDetails:
    type: object
    properties:
      name:
        type: string
      surName:
        type: string
      accounts:
        type: array
        items:
          $ref: '#/definitions/accountDetails'
  accountDetails:
    type: object
    properties:
      accountNo:
        type: string
      balance:
        type: number
      transactions:
        type: array
        items:
          type: string